[Client] Incorporate metaheader into the client, use version from transport (#387)
* Incorporate metaheader into the client, use version from transport
